+++ comp2013/src/main/java/comp2013/MyFrame.java
@@ -20,7 +20,7 @@ import javax.swing.JPanel;
  * @version Ekki viss
  */ 
 
-
+/*
 public class MyFrame extends JPanel implements KeyListener
 {
 	private static final long serialVersionUID = -3149926831770554380L;
@@ -34,9 +34,7 @@ public class MyFrame extends JPanel implements KeyListener
 
 	public void loadFrame()
 	{
-		/*
-		 * Komið í veg fyrir að myndin blikki.
-		 */
+
 		this.setDoubleBuffered(true);
 		jFrame.add(this);
 		jFrame.addKeyListener(this);
@@ -125,9 +123,7 @@ public class MyFrame extends JPanel implements KeyListener
 			this.speed_XY = 5;
 			this.length = 1;
 
-			/*
-			 * Attention : ?
-			 */
+
 			this.num = w / speed_XY;
 			newImgSnakeHead = IMG_SNAKE_HEAD;
 
@@ -156,7 +152,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = false;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, -90);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, -90);
 				}
 				break;
 
@@ -168,7 +164,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = false;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, 90);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, 90);
 				}
 				break;
 
@@ -180,7 +176,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = true;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, -180);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, -180);
 
 				}
 				break;
@@ -293,4 +289,4 @@ public class MyFrame extends JPanel implements KeyListener
 			return new Rectangle(x, y, w, h);
 		}
 	}
-}
+} */